Renamed Sampler.sample to .next()
This is due to some minor semantic issues. I am
not a huge fan of the term "Sampler", since Sample
refers to a classified point, which suggests that
Sampler produces classified points. However, a
Sampler being used to generates points for
sampling makes some sense, so removing "sample" as
the verb and replacing it with "next" to help
distinguish the purpose of the component. More
thought may be put into this for the future, but
atm this is sufficient.